improvements

+++ includes/pipe.h
@@ -4,12 +4,12 @@ using namespace std;
 class Pipe
 {
     public:
-        int speed;
-        int x;
-        vector<int> y;
-        int finishU;
-        int startB;
-        int spacing;
+        short int speed;
+        short int x;
+        vector<short int> y;
+        short int finishU;
+        short int startB;
+        short int spacing;
         void update();
         void show();
         bool isHit(Bird* bird);
